requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
click==8.1.7
Deprecated==1.2.14
fastapi==0.115.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
h11==0.14.0
idna==3.10
importlib_metadata==8.4.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
protobuf==4.25.5
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
setuptools==75.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
wrapt==1.16.0
zipp==3.20.2
